Adding Scan icon for Extract
