+++ sparkle/src/icons/stroke/RobotShared.tsx
@@ -9,16 +9,20 @@ const SvgRobotShared = (props: SVGProps<SVGSVGElement>) => (
     viewBox="0 0 24 24"
     {...props}
   >
-    <path
-      fill="currentColor"
-      d="M18 3.5a1 1 0 1 0 0-2 1 1 0 0 0 0 2ZM8.5 5.5a1.5 1.5 0 1 1-3 0 1.5 1.5 0 0 1 3 0ZM19 14.5a1 1 0 1 1-2 0 1 1 0 0 1 2 0ZM8.5 14.25c-.5.5-1.5.5-1.5.5s-1 0-1.5-.5v-.5h3v.5ZM6.25 12a.75.75 0 1 1-1.5 0 .75.75 0 0 1 1.5 0ZM8.5 12.75a.75.75 0 1 0 0-1.5.75.75 0 0 0 0 1.5Z"
-    />
-    <path
-      fill="currentColor"
-      fillRule="evenodd"
-      d="M14 6a2 2 0 0 1 2-2h4c1.105 0 2 .894 2 1.999V9a2 2 0 0 1-2 2h-3.5a2.49 2.49 0 0 1-1.103-.256c-.568-.28-1.262-.442-1.801-.11A1.25 1.25 0 0 0 13 11.699v2.24a2 2 0 0 0 1.515 1.94l.246.061c.159.04.322.06.485.06H20a2 2 0 0 1 2 2v3a2 2 0 0 1-2 2h-4a2 2 0 0 1-2-2v-1.745c0-.593-.313-1.142-.824-1.444-.648-.383-1.463-.235-2.173.017A2.995 2.995 0 0 1 10 18H4a3 3 0 0 1-3-3v-4a3 3 0 0 1 3-3h6c.49 0 .95.117 1.359.325.593.302 1.308.468 1.893.151.46-.25.748-.731.748-1.256V6ZM3 11a1 1 0 0 1 1-1h6a1 1 0 0 1 1 1v4a1 1 0 0 1-1 1H4a1 1 0 0 1-1-1v-4Zm14.25-4.25a.75.75 0 1 1-1.5 0 .75.75 0 0 1 1.5 0ZM18 9.5s1 0 1.5-.5v-.5h-3V9c.5.5 1.5.5 1.5.5Zm0 12s1 0 1.5-.5v-.5h-3v.5c.5.5 1.5.5 1.5.5Zm-1.5-2a.75.75 0 1 0 0-1.5.75.75 0 0 0 0 1.5Zm3 0a.75.75 0 1 0 0-1.5.75.75 0 0 0 0 1.5Zm.75-12.75a.75.75 0 1 1-1.5 0 .75.75 0 0 1 1.5 0Z"
-      clipRule="evenodd"
-    />
+    <g fill="currentColor" clipPath="url(#robot-shared_svg__a)">
+      <path d="M8.5 5.5a1.5 1.5 0 1 1-3 0 1.5 1.5 0 0 1 3 0Z" />
+      <path
+        fillRule="evenodd"
+        d="M14 6a2 2 0 0 1 2-2h4c1.105 0 2 .894 2 1.999V9a2 2 0 0 1-2 2h-3.5a2.49 2.49 0 0 1-1.103-.256c-.568-.28-1.262-.442-1.801-.11A1.25 1.25 0 0 0 13 11.699v2.24a2 2 0 0 0 1.515 1.94l.246.061c.159.04.322.06.485.06H20a2 2 0 0 1 2 2v3a2 2 0 0 1-2 2h-4a2 2 0 0 1-2-2v-1.745c0-.593-.313-1.142-.824-1.444-.648-.383-1.463-.235-2.173.017A2.995 2.995 0 0 1 10 18H4a3 3 0 0 1-3-3v-4a3 3 0 0 1 3-3h6c.49 0 .95.117 1.359.325.593.302 1.308.468 1.893.151.46-.25.748-.731.748-1.256V6ZM3 11a1 1 0 0 1 1-1h6a1 1 0 0 1 1 1v4a1 1 0 0 1-1 1H4a1 1 0 0 1-1-1v-4Z"
+        clipRule="evenodd"
+      />
+      <path d="M19.5 1.5a1.5 1.5 0 1 1-3 0 1.5 1.5 0 0 1 3 0ZM19.5 13.5a1.5 1.5 0 1 1-3 0 1.5 1.5 0 0 1 3 0Z" />
+    </g>
+    <defs>
+      <clipPath id="robot-shared_svg__a">
+        <path fill="#fff" d="M0 0h24v24H0z" />
+      </clipPath>
+    </defs>
   </svg>
 );
 export default SvgRobotShared;
