+++ sparkle/src/icons/stroke/index.ts
@@ -90,10 +90,10 @@ export { default as Rectangle } from "./Rectangle";
 export { default as Robot } from "./Robot";
 export { default as RobotShared } from "./RobotShared";
 export { default as Rocket } from "./Rocket";
+export { default as Scan } from "./Scan";
 export { default as Server } from "./Server";
 export { default as ShakeHands } from "./ShakeHands";
 export { default as Shapes } from "./Shapes";
-export { default as ShapesShare } from "./ShapesShare";
 export { default as Sparkles } from "./Sparkles";
 export { default as Square } from "./Square";
 export { default as Square3Stack3D } from "./Square3Stack3D";
